Correction Date et Heure de l'historique Ajout du bouton reset dans l'activité des scores Ajout d'une catégorie "Tous" non supprimable Modification de la requête de l'historique ActionBar MainJeu -> Affiche la catégorie en mode 1 joueur                                   -> Affiche le nom du joueur en mode 2 joueurs                                   -> Affiche Historique dans l'historique  Modification catégories -> impossible d'ajouter deux catégories ayant le meme nom  Modification mots -> impossible d'ajouter deux mots dans la meme catégories avec le meme nom  Modification de l'affichage du score en mode 2 joueurs  PUNTOS DE MERDE

+++ MyApplication/app/src/main/java/com/example/valentin/apppendu/Activity/GestionCategories.java
@@ -30,7 +30,7 @@ import com.example.valentin.apppendu.R;
 
 public class GestionCategories extends Activity implements View.OnClickListener, AdapterView.OnItemClickListener {
 
-    /** Boouton d'ajout d'une catégorie */
+    /** Bouton d'ajout d'une catégorie */
     private FloatingActionButton floatingActionButton;
 
     /** Bouton de retour vers le menu */
@@ -129,10 +129,10 @@ public class GestionCategories extends Activity implements View.OnClickListener,
         String nomCategorie = curseur.getString(curseur.getColumnIndex(GestionBDCategorie.CATEGORIE_NOM));
 
         // Lancement de l'activité de gestion des mots d'une catégorie
-        Intent intent = new Intent(GestionCategories.this, GestionMots.class);
-        intent.putExtra(ID_CATEGORIE, identifiant);
-        intent.putExtra(NOM_CATEGORIE, nomCategorie);
-        startActivity(intent);
+            Intent intent = new Intent(GestionCategories.this, GestionMots.class);
+            intent.putExtra(ID_CATEGORIE, identifiant);
+            intent.putExtra(NOM_CATEGORIE, nomCategorie);
+            startActivity(intent);
 
     }
 
@@ -200,11 +200,37 @@ public class GestionCategories extends Activity implements View.OnClickListener,
                     public void onClick(DialogInterface dialog, int leBouton) {
                         EditText modifSaisie = (EditText) boiteDialog.findViewById(R.id.editTextDialog);
                         String saisie = modifSaisie.getText().toString();   // Nom de la catégorie modifiée
-                        // APPEL METHODE BD (modifier)
-                        categorieModif.setLibelle(saisie);
-                        categorieDAO.updateCategorie(categorieModif);
-                        refreshListe();
-                        Toast.makeText(GestionCategories.this, "Catégorie modifiée", Toast.LENGTH_SHORT).show();
+                        // Teste pour voir si il existe déjà la catégorie
+
+                        if (!saisie.equals("")) {
+
+                            // Test pour savoir si la categorie à supprimer est la catégorie "Tous"
+                            if (categorieModif.getId() != 4) {
+                                // Mise en majuscule de la première lettre
+                                StringBuilder result = new StringBuilder(saisie);
+                                result.replace(0, 1, result.substring(0, 1).toUpperCase());
+                                saisie = result.toString();
+
+                                if (categorieDAO.recherchePresenceCategorie(saisie) == 0) {
+                                    // APPEL METHODE BD (modifier)
+                                    categorieModif.setLibelle(saisie);
+                                    categorieDAO.updateCategorie(categorieModif);
+                                    refreshListe();
+                                    Toast.makeText(GestionCategories.this, "Catégorie modifiée", Toast.LENGTH_SHORT).show();
+                                } else {
+                                    Toast.makeText(GestionCategories.this, "La catégorie " + saisie + " existe déjà", Toast.LENGTH_SHORT).show();
+                                }
+                            } else {
+                                Toast.makeText(GestionCategories.this, "Le nom de cette catégorie ne peut être modifié", Toast.LENGTH_SHORT).show();
+                            }
+
+
+                        } else {
+                            Toast.makeText(GestionCategories.this, "Le nom de la catégorie ne peut être vide", Toast.LENGTH_SHORT).show();
+                        }
+
+
+
                     }
                 });
 
@@ -253,17 +279,24 @@ public class GestionCategories extends Activity implements View.OnClickListener,
         dialogSuppression.setPositiveButton(getResources().getString(R.string.oui),
                 new DialogInterface.OnClickListener() {
                     public void onClick(DialogInterface dialog, int leBouton) {
-                        // APPEL METHODE BD (supprimer)
-                        int nbLignesSupp = 0;
-                        // TODO Suppression en cascade
-                        nbLignesSupp = categorieDAO.deleteCategorie(categorieSupprimer);
 
-                        if (nbLignesSupp >= 1) {
-                            Toast.makeText(GestionCategories.this, "\"" + categorieSupprimer.getLibelle() + "\" supprimée", Toast.LENGTH_SHORT).show();
-                            refreshListe();
+                        // Test pour savoir si la categorie à supprimer est la catégorie "Tous"
+                        if (categorieSupprimer.getId() != 4) {
+                            // APPEL METHODE BD (supprimer)
+                            int nbLignesSupp = 0;
+                            nbLignesSupp = categorieDAO.deleteCategorie(categorieSupprimer);
+
+                            if (nbLignesSupp >= 1) {
+                                Toast.makeText(GestionCategories.this, "\"" + categorieSupprimer.getLibelle() + "\" supprimée", Toast.LENGTH_SHORT).show();
+                                refreshListe();
+                            } else {
+                                Toast.makeText(GestionCategories.this, "Erreur lors de la suppression ", Toast.LENGTH_SHORT).show();
+                            }
                         } else {
-                            Toast.makeText(GestionCategories.this, "Erreur lors de la suppression ", Toast.LENGTH_SHORT).show();
+                            Toast.makeText(GestionCategories.this, "La catégorie ne peut être supprimée", Toast.LENGTH_SHORT).show();
                         }
+
+
                     }
                 });
 
@@ -297,11 +330,23 @@ public class GestionCategories extends Activity implements View.OnClickListener,
                     public void onClick(DialogInterface dialog, int leBouton) {
                         EditText ajout = (EditText) boiteDialog.findViewById(R.id.editTextAjoutCategorie);
                         String saisie = ajout.getText().toString();   // Nom de la catégorie ajoutée
+
                         if (!saisie.equals("")) {
-                            // APPEL METHODE BD (ajout)
-                            long id = categorieDAO.createCategorie(saisie);
-                            refreshListe();
-                            Toast.makeText(GestionCategories.this, "\"" + saisie + "\" a été ajoutée", Toast.LENGTH_SHORT).show();
+
+                            // Mise en majuscule de la première lettre
+                            StringBuilder result = new StringBuilder(saisie);
+                            result.replace(0, 1, result.substring(0, 1).toUpperCase());
+                            saisie = result.toString();
+
+                            if (categorieDAO.recherchePresenceCategorie(saisie) == 0) {
+                                // APPEL METHODE BD (ajout)
+                                long id = categorieDAO.createCategorie(saisie);
+                                refreshListe();
+                                Toast.makeText(GestionCategories.this, "\"" + saisie + "\" a été ajoutée", Toast.LENGTH_SHORT).show();
+                            } else {
+                                Toast.makeText(GestionCategories.this, "La catégorie " + saisie + " existe déjà", Toast.LENGTH_SHORT).show();
+                            }
+
                         } else {
                             Toast.makeText(GestionCategories.this, "Erreur lors de l'ajout de la catégorie", Toast.LENGTH_SHORT).show();
                         }
